Fix a couple of typos breaking the robot build.

+++ robot/supervisor/src/testModeController.cpp
@@ -269,12 +269,12 @@ namespace Anki {
           // Create a path and follow it
           PathFollower::AppendPathSegment_Line(0, 0.0, 0.0, 300, -300,
                                                PF_TARGET_SPEED_MMPS, PF_ACCEL_MMPS2, PF_DECEL_MMPS2);
-          float arc1_radius = sqrt(5000);  // Radius of sqrt(50^2 + 50^2)
+          float arc1_radius = sqrt((float)5000);  // Radius of sqrt(50^2 + 50^2)
           PathFollower::AppendPathSegment_Arc(0, 350, -250, arc1_radius, -0.75*PI, 0.75*PI,
                                               PF_TARGET_SPEED_MMPS, PF_ACCEL_MMPS2, PF_DECEL_MMPS2);
           PathFollower::AppendPathSegment_Line(0, 350 + arc1_radius, -250, 350 + arc1_radius, 200,
                                                PF_TARGET_SPEED_MMPS, PF_ACCEL_MMPS2, PF_DECEL_MMPS2);
-          float arc2_radius = sqrt(20000); // Radius of sqrt(100^2 + 100^2)
+          float arc2_radius = sqrt((float)20000); // Radius of sqrt(100^2 + 100^2)
           //PathFollower::AppendPathSegment_Arc(0, 0.35 + arc1_radius - arc2_radius, 0.2, arc2_radius, 0, PIDIV2);
           PathFollower::AppendPathSegment_Arc(0, 350 + arc1_radius - arc2_radius, 200, arc2_radius, 0, 3*PIDIV2,
                                               PF_TARGET_SPEED_MMPS, PF_ACCEL_MMPS2, PF_DECEL_MMPS2);
